refactor: id types in entities are reference type

+++ common-codebase/src/main/java/dev/faruk/commoncodebase/dto/OfferDTO.java
@@ -13,9 +13,9 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class OfferDTO {
-    private int id;
+    private Long id;
     private String name;
-    private float discount;
+    private Double discount;
     private String validUntil;
     private String validSince;
     private List<OfferProductDTO> requiredProducts;
